clean up utils
